When building the package, we can publish dotnet artefacts in parallel procs, and we can run IO-bound compression in parallel threads.

time python build_differ.py measures roughly a 30-35 second decrease in execution time on my new MBP. 

Before:
Build and compression complete.
python build_differ.py  41.88s user 1.93s system 101% cpu 43.191 total

Now:
python build_differ.py  53.09s user 3.63s system 484% cpu 11.706 total