@ross-mcnairn-dev 2025-05-13
Hey @JSv4 let's drop this - we've noticed sporadic build failures that seem to be caused by a race condition. I haven't had time to properly debug, but for now we're pointing back to the repo here and accepting the higher build times.